3-state model (wider distance)
Data stream now considers the closest 15 points rather than the closest 5. Works a little better in some ways, but also lumps in nearby nesting attempts. Also means that we'll need 15 points on the ground before we can determine mortality, but that was probably always going to be the case. Model works pretty well on the subsetted data: more data streams? Apply to full dataset?